feat: fix schema, ObjectId and other small bugs
