+++ src/app.module.ts
@@ -9,7 +9,7 @@ import { UserModule } from "./users/user.module";
 import { AdminModule } from "./admin/admin.module";
 import { PriceModule } from "./price/price.module";
 import { DesignModule } from "./designs/design.module";
-import { CartItemModule } from "./cart_item/cart_item.module";
+import { CartItemModule } from "./cart_items/cartItem.module";
 
 @Module({
   imports: [
