+++ src/cart_items/cartItem.testData.ts
@@ -1,0 +1,414 @@
+export const mockCartItem = {
+  user_id: "user123",
+  design: {
+    designName: "Court Canva 1",
+    tileColor: [
+      {
+        location: "threePoint",
+        color: "#7088B1",
+      },
+      {
+        location: "courtArea",
+        color: "#E18E11",
+      },
+      {
+        location: "topKeyArea",
+        color: "#B6B6B6",
+      },
+      {
+        location: "border",
+        color: "#834085",
+      },
+      {
+        location: "keyArea",
+        color: "#2C4E8A",
+      },
+      {
+        location: "circleArea",
+        color: "#B6B6B6",
+      },
+    ],
+    courtSize: {
+      name: "Pro Full Court",
+      length: 28000,
+      width: 15000,
+      threePointLine: 900,
+      threePointRadius: 6600,
+      centreCircleRadius: 1800,
+      restrictedAreaLength: 5790,
+      restrictedAreaWidth: 4800,
+      sideBorderWidth: 1000,
+      lengthOfCorner: 1575,
+      lineBorderWidth: 200,
+      designName: "Court Canva 1",
+    },
+  },
+  quotation: "67445.40",
+  quotationDetails: [
+    {
+      color: "#7088B1",
+      quantity: 1273,
+    },
+    {
+      color: "#E18E11",
+      quantity: 2572,
+    },
+    {
+      color: "#B6B6B6",
+      quantity: 231,
+    },
+    {
+      color: "#834085",
+      quantity: 1000,
+    },
+    {
+      color: "#2C4E8A",
+      quantity: 624,
+    },
+  ],
+};
+
+export const mockCartItemArray = [
+  {
+    user_id: "user123",
+    design: {
+      designName: "Court Canva 1",
+      tileColor: [
+        {
+          location: "threePoint",
+          color: "#7088B1",
+        },
+        {
+          location: "courtArea",
+          color: "#E18E11",
+        },
+        {
+          location: "topKeyArea",
+          color: "#B6B6B6",
+        },
+        {
+          location: "border",
+          color: "#834085",
+        },
+        {
+          location: "keyArea",
+          color: "#2C4E8A",
+        },
+        {
+          location: "circleArea",
+          color: "#B6B6B6",
+        },
+      ],
+      courtSize: {
+        name: "Pro Full Court",
+        length: 28000,
+        width: 15000,
+        threePointLine: 900,
+        threePointRadius: 6600,
+        centreCircleRadius: 1800,
+        restrictedAreaLength: 5790,
+        restrictedAreaWidth: 4800,
+        sideBorderWidth: 1000,
+        lengthOfCorner: 1575,
+        lineBorderWidth: 200,
+        designName: "Court Canva 1",
+      },
+    },
+    quotation: "67445.40",
+    quotationDetails: [
+      {
+        color: "#7088B1",
+        quantity: 1273,
+      },
+      {
+        color: "#E18E11",
+        quantity: 2572,
+      },
+      {
+        color: "#B6B6B6",
+        quantity: 231,
+      },
+      {
+        color: "#834085",
+        quantity: 1000,
+      },
+      {
+        color: "#2C4E8A",
+        quantity: 624,
+      },
+    ],
+  },
+  {
+    user_id: "user123",
+    design: {
+      designName: "Court Canva 2",
+      tileColor: [
+        {
+          location: "threePoint",
+          color: "#44BC84",
+        },
+        {
+          location: "courtArea",
+          color: "#558BE7",
+        },
+        {
+          location: "topKeyArea",
+          color: "#EB4335",
+        },
+        {
+          location: "border",
+          color: "#834085",
+        },
+        {
+          location: "keyArea",
+          color: "#2C4E8A",
+        },
+        {
+          location: "circleArea",
+          color: "#EB4335",
+        },
+      ],
+      courtSize: {
+        name: "Pro Half Court",
+        length: 14000,
+        width: 15000,
+        threePointLine: 900,
+        threePointRadius: 6600,
+        centreCircleRadius: 1800,
+        restrictedAreaLength: 5790,
+        restrictedAreaWidth: 4800,
+        sideBorderWidth: 0,
+        lengthOfCorner: 1575,
+        lineBorderWidth: 200,
+        designName: "Court Canva 1",
+      },
+    },
+    quotation: "27782.70",
+    quotationDetails: [
+      {
+        color: "#44BC84",
+        quantity: 632,
+      },
+      {
+        color: "#558BE7",
+        quantity: 1295,
+      },
+      {
+        color: "#EB4335",
+        quantity: 119,
+      },
+      {
+        color: "#834085",
+        quantity: 0,
+      },
+      {
+        color: "#2C4E8A",
+        quantity: 304,
+      },
+    ],
+  },
+  {
+    user_id: "user123",
+    design: {
+      designName: "Court Canva 3",
+      tileColor: [
+        {
+          location: "threePoint",
+          color: "#FBBC05",
+        },
+        {
+          location: "courtArea",
+          color: "#EBD935",
+        },
+        {
+          location: "topKeyArea",
+          color: "#EB4335",
+        },
+        {
+          location: "border",
+          color: "#834085",
+        },
+        {
+          location: "keyArea",
+          color: "#558BE7",
+        },
+        {
+          location: "circleArea",
+          color: "#EB4335",
+        },
+      ],
+      courtSize: {
+        name: "Medium Court",
+        length: 10000,
+        width: 7000,
+        threePointLine: 900,
+        threePointRadius: 6600,
+        centreCircleRadius: 1800,
+        restrictedAreaLength: 5790,
+        restrictedAreaWidth: 4800,
+        sideBorderWidth: 0,
+        lengthOfCorner: 1575,
+        lineBorderWidth: 200,
+        designName: "Court Canva 1",
+      },
+    },
+    quotation: "9260.90",
+    quotationDetails: [
+      {
+        color: "#FBBC05",
+        quantity: 263,
+      },
+      {
+        color: "#EBD935",
+        quantity: 189,
+      },
+      {
+        color: "#EB4335",
+        quantity: 60,
+      },
+      {
+        color: "#834085",
+        quantity: 0,
+      },
+      {
+        color: "#558BE7",
+        quantity: 304,
+      },
+    ],
+  },
+];
+
+export const mockNewCartItem = {
+  user_id: "user123",
+  design: {
+    designName: "Court Canva 1",
+    tileColor: [
+      {
+        location: "threePoint",
+        color: "#7088B1",
+      },
+      {
+        location: "courtArea",
+        color: "#E18E11",
+      },
+      {
+        location: "topKeyArea",
+        color: "#B6B6B6",
+      },
+      {
+        location: "border",
+        color: "#834085",
+      },
+      {
+        location: "keyArea",
+        color: "#2C4E8A",
+      },
+      {
+        location: "circleArea",
+        color: "#B6B6B6",
+      },
+    ],
+    courtSize: {
+      name: "Pro Full Court",
+      length: 28000,
+      width: 15000,
+      threePointLine: 900,
+      threePointRadius: 6600,
+      centreCircleRadius: 1800,
+      restrictedAreaLength: 5790,
+      restrictedAreaWidth: 4800,
+      sideBorderWidth: 1000,
+      lengthOfCorner: 1575,
+      lineBorderWidth: 200,
+      designName: "Court Canva 1",
+    },
+  },
+  quotation: "67445.40",
+  quotationDetails: [
+    {
+      color: "#7088B1",
+      quantity: 1273,
+    },
+    {
+      color: "#E18E11",
+      quantity: 2572,
+    },
+    {
+      color: "#B6B6B6",
+      quantity: 231,
+    },
+    {
+      color: "#834085",
+      quantity: 1000,
+    },
+    {
+      color: "#2C4E8A",
+      quantity: 624,
+    },
+  ],
+  createdAt: new Date(),
+  updatedAt: new Date(),
+  image: "url12345566721",
+};
+
+export const mockCartItemInDatabase = {
+  isExpired: false,
+  _id: "6320bd57f3dee2ee6deeecf2",
+  user_id: "user123",
+  designName: "CourtCanvas1.a",
+  tileColor: [
+    {
+      location: "threePoint",
+      color: "#72818B",
+    },
+    {
+      location: "courtArea",
+      color: "#B61313",
+    },
+    {
+      location: "topKeyArea",
+      color: "#B61313",
+    },
+    {
+      location: "border",
+      color: "#195955",
+    },
+    {
+      location: "keyArea",
+      color: "#2C4E8A",
+    },
+    {
+      location: "circleArea",
+      color: "#606F14",
+    },
+  ],
+  courtSize: {
+    name: "Pro Full Court",
+    length: 28000,
+    width: 15000,
+    centreCircleRadius: 1800,
+    threePointRadius: 6600,
+    threePointLine: 900,
+    lengthOfCorner: 1575,
+    restrictedAreaLength: 5790,
+    restrictedAreaWidth: 4900,
+    sideBorderWidth: 1000,
+    lineBorderWidth: 200,
+  },
+  createdAt: "2022-09-13T17:26:43.520Z",
+  updatedAt: "2022-09-14T10:39:11.476Z",
+  isDeleted: false,
+  quotation: "$AU132354123",
+  image: "url12345566778aaaaaa",
+  quotationDetails: [
+    {
+      color: "red",
+      quantity: 5,
+    },
+    {
+      color: "blue",
+      quantity: 20,
+    },
+  ],
+  __v: 0,
+};
