+++ src/cart_items/dto/update-cartItem.dto.ts
@@ -1,6 +1,6 @@
 import { PartialType } from "@nestjs/mapped-types";
 import { Exclude } from "class-transformer";
-import { CreateCartItemDto } from "./create-cart_item.dto";
+import { CreateCartItemDto } from "./create-cartItem.dto";
 
 export class UpdateCartItemDto extends PartialType(CreateCartItemDto) {
   @Exclude()
